More shop work

+++ RazzleServer.DB/ShopRechargeEntity.cs
@@ -6,7 +6,7 @@ namespace RazzleServer.Data
     {
         [Key]
         public int Id { get; set; }
-        public int TierId { get; set; }
+        public byte TierId { get; set; }
         public int ItemId { get; set; }
         public double Price { get; set; }
     }
